Manual delimiter specification
We received a file from a client with the below format.

    field1~ field2~ field3
    value1~1~1
    value2~2~2

At present, data-lake-accelerator will fail to parse this file when defining a new feed (found at the path feed-mgr/index.html#/define-feed) as core/schema-discovery/schema-discovery-parser/src/main/java/com/thinkbiganalytics/schema/TextFileParser.java:guessDelimiter() is unable to work out which delimiter is in use.

This pull request will allow users to manually specify which delimiter they wish to dilineate fields when defining the schema for a table for a feed. Without this, users can only receive an error message, "An unexpected error occurred An unknown error has occurred. This is very likely due to invalid input" along with a HTTP 500 when attempting to send the file to the v1/feedmgr/feeds/table/sample-file API.