Remove localhost broker from activemq.properties
This fixes a bug where nifi can't be configured to use the correct activemq broker url since this file overwrites the values in ext-config. 
>In: https://github.com/Teradata/kylo/blob/master/plugins/jms-service-activemq/src/main/java/com/thinkbiganalytics/jms/activemq/ActiveMqConfig.java

we see that "file:${kylo.nifi.configPath}/config.properties" is defined before "classpath:activemq.properties". activemq.properties is on the classpath and contains the setting for localhost for activemq which overwrites the value we specify in ext-config. 

https://github.com/Teradata/kylo/blob/master/plugins/jms-service-activemq/src/main/resources/activemq.properties